requirements.txt
customtkinter
pyotp
keyring
Pillow
cryptography
pyperclip